### 2023-08-21 #5775 @editorialbot
[PRE REVIEW]: CL-PFU: A suite of R packages for energy conversion chain analysis
**Submitting author:** @MatthewHeun (<a href="http://orcid.org/0000-0002-7438-214X">Matthew Heun</a>)
**Repository:** https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper
**Branch with paper.md** (empty if default branch):
**Version:** This submission covers several packages, each with its own current version.
