### 2023-08-27 #5775 @MatthewHeun
I looked in the list of reviewers for the combination of R knowledge and keyword "energy." From that list, Nick Gorman (nick-gorman), Kalai Ramea (kramea), and Neeraj Bokde (neerajdhanraj) stand out as good referee candidates. Nick Gorman and Kalai Ramea have expertise in energy and climate, the application area of the software described in this paper. Neeraj Bodke has expertise in data cleaning, one of the important functions of the calculation pipelines available in the software described in this paper.
