+++ paper.bib
@@ -15,8 +15,7 @@
 	note = {R package version 0.1.1},
 	title = {EROITools: Tools for calculating and aggregating Energy Return On Investment values},
 	url = {https://doi.org/10.5281/zenodo.7696869},
-	year = {2022},
-	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5819143}}
+	year = {2022}}
 
 @manual{Aramendia:2022aa,
 	author = {Emmanuel Aramendia and Matthew Kuperus Heun},
@@ -25,8 +24,7 @@
 	note = {R package version 0.1.5},
 	title = {ECCTools: Tools for modifying the Energy Conversion Chain},
 	url = {https://doi.org/10.5281/zenodo.5841962},
-	year = {2022},
-	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5819143}}
+	year = {2022}}
 
 @phdthesis{Aramendia:2023aa,
 	author = {Emmanuel Aramendia and Paul Brockway and Peter Taylor and J. Norman and Matthew Kuperus Heun and Zeke Marshall},
@@ -35,6 +33,7 @@
 	school = {University of Leeds},
 	title = {Chapter 4: Estimation of fossil fuels useful stage energy return on investment and implications for renewable energy systems \normalfont in {E}xploring the useful energy implications of the global energy transition: a net energy perspective},
 	year = {2023}}
+
 
 @misc{International-Energy-Agency:2023aa,
 	author = {{International Energy Agency}},
@@ -45,67 +44,6 @@
 	title = {{IEA} {W}orld {E}xtended {E}nergy {B}alances},
 	year = {2023}}
 
-@article{Pearson:2017,
-	archiveprefix = {arXiv},
-	author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-	eprint = {1703.04627},
-	journal = {ArXiv e-prints},
-	keywords = {Astrophysics - Astrophysics of Galaxies},
-	month = mar,
-	title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-	url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-	year = 2017,
-	bdsk-url-1 = {http://adsabs.harvard.edu/abs/2017arXiv170304627P}}
-
-@book{Binney:2008,
-	author = {{Binney}, J. and {Tremaine}, S.},
-	booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
-	publisher = {Princeton University Press},
-	title = {{Galactic Dynamics: Second Edition}},
-	url = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-	year = 2008,
-	bdsk-url-1 = {http://adsabs.harvard.edu/abs/2008gady.book.....B}}
-
-@article{gaia,
-	archiveprefix = {arXiv},
-	author = {{Gaia Collaboration}},
-	doi = {10.1051/0004-6361/201629272},
-	eprint = {1609.04153},
-	journal = {Astronomy and Astrophysics},
-	keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
-	month = nov,
-	primaryclass = {astro-ph.IM},
-	title = {{The Gaia mission}},
-	url = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
-	volume = 595,
-	year = 2016,
[…]
