+++ paper.bib
@@ -65,7 +65,9 @@
 	pages = {219-273},
 	title = {Exergy, power, and work in the {US} economy, 1900--1998},
 	volume = {28},
-	year = {2003}}
+	year = {2003}, 
+	doi = {10.1016/s0360-5442(02)00089-0}}
+
 
 @article{Brockway:2014aa,
 	author = {Paul E. Brockway and John R. Barrett and Timothy J. Foxon and Julia K. Steinberger},
@@ -73,7 +75,9 @@
 	pages = {9874-9881},
 	title = {Divergence of Trends in {US} and {UK} Aggregate Exergy Efficiencies 1960--2010},
 	volume = {48},
-	year = {2014}}
+	year = {2014}, 
+	doi = {10.1021/es501217t}}
+
 
 @article{Brockway:2015aa,
 	author = {Paul E. Brockway and Julia K. Steinberger and John R. Barrett and Timothy J. Foxon},
@@ -81,7 +85,8 @@
 	pages = {892-903},
 	title = {Understanding {C}hina's past and future energy demand: An exergy efficiency and decomposition analysis},
 	volume = {155},
-	year = {2015}}
+	year = {2015}, 
+	doi = {10.1016/j.apenergy.2015.05.082}}
 	
 
 @article{Common:1976wn,
@@ -92,7 +97,8 @@
 	pages = {158-165},
 	title = {The economics of energy analysis reconsidered},
 	volume = {4},
-	year = {1976}}
+	year = {1976}, 
+	doi = {10.1016/0301-4215(76)90008-2}}
 
 
 @techreport{De-Stercke:2014,
@@ -112,7 +118,8 @@
 	pages = {253-270},
 	title = {Society exergy analysis: a comparison of different societies},
 	volume = {26},
-	year = {2001}}
+	year = {2001}, 
+	doi = {10.1016/s0360-5442(00)00070-0}}
 
 
 @article{Guevara:2017,
@@ -121,7 +128,9 @@
 	pages = {261-269},
 	title = {The multi-factor energy input--output model},
 	volume = {61},
-	year = {2017}}
+	year = {2017}, 
+	doi = {10.1016/j.eneco.2016.11.020}}
+
 
 @article{Heun:2018,
 	author = {Matthew Kuperus Heun and Anne Owen and Paul E. Brockway},
@@ -130,7 +139,9 @@
 	pages = {1134-1162},
 	title = {A physical supply-use table framework for energy analysis on the energy conversion chain},
 	volume = {226},
-	year = {2018}}
+	year = {2018}, 
+	doi = {10.1016/j.apenergy.2018.05.109}}
+
 
 @manual{Heun-RCLabels:2023,
 	author = {Matthew Kuperus Heun},
@@ -138,6 +149,7 @@
 	title = {RCLabels: Tools for munging and manipulating row and column labels on matrices},
 	url = {https://doi.org/10.5281/zenodo.5819143},
 	year = {2023}}
+
 
 @manual{Heun-matsbyname:2023,
 	author = {Matthew Kuperus Heun},
@@ -218,9 +230,9 @@
 	number = {57},
 	pages = {2959},
 	title = {The targets {R} package: A dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing},
-	url = {https://doi.org/10.21105/joss.02959},
 	volume = {6},
-	year = {2021}}
+	year = {2021}, 
+	doi = {10.21105/joss.02959}}
 
 
 @dataset{Marshall:2023aa,
@@ -253,7 +265,8 @@
 	pages = {704-715},
[…]
