### 2023-08-27 #5775 @MatthewHeun
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1051/0004-6361/201629272 is OK
> - 10.1051/0004-6361/201322068 is OK
> - 10.5281/zenodo.5228359 is OK
> - 10.21105/joss.01686 is OK
>
> MISSING DOIs
>
> - 10.1016/j.apenergy.2021.118413 may be a valid DOI for title: Developing a Multi-Regional Physical Supply Use Table framework to improve the accuracy and reliability of energy analysis
> - 10.1016/s0360-5442(02)00089-0 may be a valid DOI for title: Exergy, power, and work in the US economy, 1900–1998
> - 10.1021/es501217t may be a valid DOI for title: Divergence of Trends in US and UK Aggregate Exergy Efficiencies 1960–2010
> - 10.1016/j.apenergy.2015.05.082 may be a valid DOI for title: Understanding China’s past and future energy demand: An exergy efficiency and decomposition analysis
> - 10.1016/0301-4215(76)90008-2 may be a valid DOI for title: The economics of energy analysis reconsidered
> - 10.1016/s0360-5442(00)00070-0 may be a valid DOI for title: Society exergy analysis: a comparison of different societies
> - 10.1016/j.eneco.2016.11.020 may be a valid DOI for title: The multi-factor energy input–output model
> - 10.1016/j.apenergy.2018.05.109 may be a valid DOI for title: A physical supply-use table framework for energy analysis on the energy conversion chain
> - 10.21105/joss.02959 may be a valid DOI for title: The targets R package: A dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing
> - 10.1016/j.energy.2014.08.068 may be a valid DOI for title: Decomposition of useful work intensity: The EU (European Union)-15 countries from 1960 to 2009
> - 10.1007/s41247-022-00096-z may be a valid DOI for title: The Contributions of Muscle and Machine Work to Land and Labor Productivity in World Agriculture Since 1800
>
> INVALID DOIs
>
> - 10.5518/1199 is INVALID

I have added the above DOIs to the paper.

> - 10.5518/1199 is INVALID

This DOI is not yet valid, because we are waiting to mint this DOI with the University of Leeds. The DOI has been assigned, though.
