+++ paper.Rmd
@@ -86,7 +86,7 @@ output:
 [link-eroitools]:        https://github.com/earamendia/EROITools
 [link-iea]:              https://www.iea.org
 [link-ieatools]:         https://github.com/MatthewHeun/IEATools
-[link-ieaweb]:           https://www.iea.org/data-and-statistics/data-product/world-energy-balances
+[link-ieaeweb]:          https://www.iea.org/data-and-statistics/data-product/world-energy-balances
 [link-ilo]:              https://www.ilo.org/
 [link-fao]:              https://www.fao.org/
 [link-matsbyname]:       https://github.com/MatthewHeun/matsbyname
@@ -155,7 +155,7 @@ analyzed the ECCs of individual countries
 using linked spreadsheets,
 often starting with primary- and final-stage data 
 from the [International Energy Agency][link-iea]'s (IEA's)
-[world energy balances][link-ieaweb] (WEBs)
+[extended world energy balances][link-ieaeweb] (EWEBs)
 [@Ayres:2003ec; @Serrenho:2014aa; @Brockway:2014aa; @Brockway:2015aa]. 
 Data were stored in varying and inconsistent formats.
 An early SEA database [@De-Stercke:2014]
@@ -191,7 +191,7 @@ processing stages
 and, ultimately, 
 to final demand
 in a way that is compatible with the [IEA][link-iea]'s
-[WEB][link-ieaweb] data
+[EWEB][link-ieaeweb] data
 [@International-Energy-Agency:2023aa].
 We chose the PSUT framework as the data format for the new `R` packages, 
 because it succinctly describes an entire ECC
@@ -247,7 +247,7 @@ Broadly speaking,
 four calculation steps are required to create a CL-PFU database. 
 Each step is assisted by functions in one or more of the new `R` packages.
 First, the IEA's primary- and final-stage 
-WEB data must be converted
+EWEB data must be converted
 to RUVY matrices for each country and year, 
 a task accomplished by functions in the [IEATools][link-ieatools] package [@Heun-IEATools:2023]. 
 Second, human and animal muscle work must be calculated from 
@@ -322,7 +322,7 @@ The packages in the following table are available on GitHub.
 
 | Package | Purpose |
 |:--------|:--------|
-| [IEATools][link-ieatools]             | Converts [IEA WEB][link-ieaweb] data to RUVY matrices in [matsindf][link-matsindf] data frames |
+| [IEATools][link-ieatools]             | Converts [IEA EWEB][link-ieaeweb] data to RUVY matrices in [matsindf][link-matsindf] data frames |
 | [MWTools][link-mwtools]               | Converts [ILO][link-ilo] and [FAO][link-fao] data to RUVY matrices of human and animal muscle work in [matsindf][link-matsindf] data frames |
 | [Recca][link-recca]                   | Performs `R` energy conversion chain analysis |
 | [PFUSetup][link-pfusetup]             | Identifies input and output locations for the [PFUDatabase][link-pfudatabase] and [PFUAggDatabase][link-pfuaggdatabase] pipelines |
@@ -333,11 +333,11 @@ The packages in the following table are available on GitHub.
 Input data for the PFU database can be found in @Marshall:2023aa.
 Access to the full CL-PFU database can be obtained via correspondence 
 with author [PEB](mailto:P.E.Brockway@leeds.ac.uk)^[Because the PFU Database contains
-primary and final energy [IEA WEB][link-ieaweb] data, 
+primary and final energy [IEA EWEB][link-ieaeweb] data, 
 use of the PFU Database
-is restricted to those who have access to [IEA][link-iea] [WEB][link-ieaweb] data.].
[…]
