### 2023-11-13 #6057 @editorialbot
[REVIEW]: CL-PFU: A suite of R packages for energy conversion chain analysis
**Submitting author:** @MatthewHeun (<a href="http://orcid.org/0000-0002-7438-214X">Matthew Heun</a>)
**Repository:** https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.9
