### 2023-11-26 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#1 @nmstreethran
JOSS Review
Hi @MatthewHeun, I just completed reviewing your JOSS submission (https://github.com/openjournals/joss-reviews/issues/6057). I have some comments on the documentation and software paper which you can find below. Please let me know if anything is unclear. Many thanks.

## Documentation

I think users would benefit from a separate documentation (e.g. webpage or README) for the CL-PFU suite of packages. Based on my JOSS checklist, the following items must be included in the documentation:

- Statement of need
- Installation instructions
- Example usage
- Automated tests
- Functionality documentation
- Community guidelines

Some of these (statement of need, example usage) are already in the paper, so you'll only have to copy and paste them. I did run into some issues running the example though (see details under the **Software Paper** heading below). I would also recommend adding the tables listing the packages as this will link to their GitHub repositories which contain the functionality documentation. Perhaps you could create a README in the [EnergyEconomyDecoupling/CL-PFU-JOSS-Paper](https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper) repository and add these?

The [EnergyEconomyDecoupling/PFUAggDatabase](https://github.com/EnergyEconomyDecoupling/PFUAggDatabase) repository's README has a mistake; it says `PFUAggWorkflow` instead of `PFUAggDatabase`.

I couldn't find any community guidelines in the package repositories. JOSS requires "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". Tests are present in the package repositories but brief instructions on how to run them in the community guidelines would be useful for those wishing to contribute to the software.

I think it would be great if you can include the command to install the entire suite in the documentation. I used the following command to install all dependencies. I also added the `MatthewHeun/MKHthemes` package as I encountered the following error during installation: `ERROR: dependency ‘MKHthemes’ is not available for package ‘PFUDatabase’`.

```r
# install devtools if not already installed
# install.packages("devtools")

# install GitHub packages
devtools::install_github(c("MatthewHeun/IEATools",
                           "EnergyEconomyDecoupling/MWTools",
                           "MatthewHeun/Recca",
                           "EnergyEconomyDecoupling/PFUSetup",
                           "EnergyEconomyDecoupling/PFUPipelineTools",
                           "MatthewHeun/MKHthemes",
                           "EnergyEconomyDecoupling/PFUDatabase",
                           "EnergyEconomyDecoupling/PFUAggDatabase"))
```

## Software Paper

The paper is well written with some minor formatting issues. The `ℹ` in the example code (line 108 in the PDF) seems to be missing. You might want to consider breaking (3) and (4) in the example into multiple lines so they don't overflow in the right margin.

As mentioned above, I tried running the example but ran into a few issues. I couldn't run (2) without installing `FAOSTAT` and `Rilostat`. I also had to install `future.callr` to run (3). Should these be part of the dependencies? I also ran into this error when running the PFUDatabase pipeline in (3) and got stuck there. Any idea why this is happening?

```
...
✖ error target AMWPFUDataRaw
▶ end pipeline [11.248 seconds]
Error:
! Error running targets::tar_make_future()
  Error messages: targets::tar_meta(fields = error, complete_only = TRUE)
  Debugging guide: https://books.ropensci.org/targets/debugging.html
  How to ask for help: https://books.ropensci.org/targets/help.html
  Last error: Can't recycle `Year` (size 61) to size 0.
```
