### 2023-11-27 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#1 @MatthewHeun
@nmstreethran Thank you for your review! I appreciate your work. I will make separate GitHub issues for each item in your review so that I can track them individually. Doing so will also separate any discussion of each item, too. When I have built issues for each of your items, I will close this issue.

Thanks again for your good work!
