### 2023-11-27 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#1 @MatthewHeun
I moved all of these comments to separate issues #2 - #8, so they can be tracked individually. Thanks for your nice work on the review!
