### 2023-11-27 faf6728
Split lines to avoid bleeding into margin. Addresses #7.
