### 2023-11-28 e7a4f86
head(as.data.frame(psut)) to avoid nonprinting "i"
