### 2023-12-10 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#9 @abhishektiwari
Fix dual licensing 
Repositions seems to have dual licences (`LICENSE` vs, `LICENSE.md`). Given JOSS only accepts OSI approved software licence, please fix licence for following packages.

- [PFUAggDatabase](https://github.com/EnergyEconomyDecoupling/PFUAggDatabase)
- [PFUDatabase](https://github.com/EnergyEconomyDecoupling/PFUDatabase)
- [PFUPipelineTools](https://github.com/EnergyEconomyDecoupling/PFUPipelineTools)
- [PFUSetup](https://github.com/EnergyEconomyDecoupling/PFUSetup)
- [Recca](https://github.com/MatthewHeun/Recca)
- [MWTools](https://github.com/EnergyEconomyDecoupling/MWTools)
- [IEATools](https://github.com/MatthewHeun/IEATools)
- [RCLabels](https://github.com/MatthewHeun/RCLabels)
- [matsbyname](https://github.com/MatthewHeun/matsbyname)
- [matsindf](https://github.com/MatthewHeun/matsindf)

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6057)
