### 2023-12-10 #6057 @abhishektiwari
@MatthewHeun Please see initial feedback.

# General checks
## Repository
 Is the source code for this software available at the https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper? **No**.

Source code for packages in CL-PFU suite is available on different GitHub repositories, but there is no direct way to browse them from the [GitHub repo used in this submission](https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper). Given this repo will be linked to digital version of paper, I strongly recommend authors to add a `README.md` to repo pointing to following 10 packages covered by the paper and under the scope of review.

- [PFUAggDatabase](https://github.com/EnergyEconomyDecoupling/PFUAggDatabase)
- [PFUDatabase](https://github.com/EnergyEconomyDecoupling/PFUDatabase)
- [PFUPipelineTools](https://github.com/EnergyEconomyDecoupling/PFUPipelineTools)
- [PFUSetup](https://github.com/EnergyEconomyDecoupling/PFUSetup)
- [Recca](https://github.com/MatthewHeun/Recca)
- [MWTools](https://github.com/EnergyEconomyDecoupling/MWTools)
- [IEATools](https://github.com/MatthewHeun/IEATools)
- [RCLabels](https://github.com/MatthewHeun/RCLabels)
- [matsbyname](https://github.com/MatthewHeun/matsbyname)
- [matsindf](https://github.com/MatthewHeun/matsindf)

## Licence
Repositions seems to have dual licences (`LICENSE` vs, `LICENSE.md`). Given JOSS only accepts OSI approved software licence, please [fix licence](https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper/issues/9) for following packages.

- [PFUAggDatabase](https://github.com/EnergyEconomyDecoupling/PFUAggDatabase)
- [PFUDatabase](https://github.com/EnergyEconomyDecoupling/PFUDatabase)
- [PFUPipelineTools](https://github.com/EnergyEconomyDecoupling/PFUPipelineTools)
- [PFUSetup](https://github.com/EnergyEconomyDecoupling/PFUSetup)
- [Recca](https://github.com/MatthewHeun/Recca)
- [MWTools](https://github.com/EnergyEconomyDecoupling/MWTools)
- [IEATools](https://github.com/MatthewHeun/IEATools)
- [RCLabels](https://github.com/MatthewHeun/RCLabels)
- [matsbyname](https://github.com/MatthewHeun/matsbyname)
- [matsindf](https://github.com/MatthewHeun/matsindf)

# Documentation

Please add `A statement of need` and/or `Community guidelines` for following packages.

- [PFUAggDatabase](https://github.com/EnergyEconomyDecoupling/PFUAggDatabase)
- [PFUDatabase](https://github.com/EnergyEconomyDecoupling/PFUDatabase)
- [PFUPipelineTools](https://github.com/EnergyEconomyDecoupling/PFUPipelineTools)
- [PFUSetup](https://github.com/EnergyEconomyDecoupling/PFUSetup)
- [Recca](https://github.com/MatthewHeun/Recca)
- [MWTools](https://github.com/EnergyEconomyDecoupling/MWTools)
- [IEATools](https://github.com/MatthewHeun/IEATools)
- [RCLabels](https://github.com/MatthewHeun/RCLabels)
- [matsbyname](https://github.com/MatthewHeun/matsbyname)
- [matsindf](https://github.com/MatthewHeun/matsindf)

# Paper

## Summary
Given the large number of packages (10) covered under paper and CL-PFU suite, paper summary can be updated to reflect two different types of packages included in the suite

— Utility packages in support of PSUT framework (3)
— Utility package in support of CL-PFU database (7)
