+++ References.bib
@@ -1,0 +1,21 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+  
+  
+  %% Created for Matt at 2018-10-31 20:55:17 -0400 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{Heun:2018,
+  Author = {Matthew Kuperus Heun and Anne Owen and Paul E. Brockway},
+  Date-Modified = {2018-10-31 20:55:16 -0400},
+  Doi = {10.1016/j.apenergy.2018.05.109},
+  Journal = {Applied Energy},
+  Month = {Sep},
+  Pages = {1134-1162},
+  Title = {A physical supply-use table framework for energy analysis on the energy conversion chain},
+  Volume = {226},
+  Year = {2018}}
