### 2023-12-11 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#9 @MatthewHeun
There are two pieces to this comment: dual licenses and OSI approval.

### OSI approval

All software in this submission carries the MIT license, because we wanted a permissive license so people can use our code with minimal restrictions. Unless I'm missing something,

https://opensource.org/license/MIT/

indicates that the MIT license is approved by OSI. So unless I'm missing something, our submission is OK regarding OSI approval.

### Dual licenses

There is actually only a single license (MIT) that uses two files (LICENSE and LICENSE.md). We follow best practices for R packages described here:

https://r-pkgs.org/license.html#key-files

and implemented in the [usethis](https://usethis.r-lib.org) package. We used the function [usethis::use_mit_license()](https://usethis.r-lib.org/reference/licenses.html) to create the license files for all packages under this submission.

Again, unless I'm missing something, our submission is OK regarding license files.

At this point, I am closing this issue. But if @abhishektiwari sees something that I'm missing, I'll be happy to re-open and address what I don't currently understand.
