+++ README.md
@@ -9,15 +9,34 @@
 
 This repository contains a paper submitted to the [Journal of Open
 Source Software (JOSS)](https://joss.theoj.org) covering the suite of
-`R` packages associated with the CL-PFU database[^1].
+`R` packages associated with the CL-PFU[^1] database.
+
+## Statement of need
+
+Historically, Societal Exergy Analysis (SEA) practitioners have analyzed
+the ECCs of individual countries using linked spreadsheets, often
+starting with primary- and final-stage data from the [International
+Energy Agency](https://www.iea.org)’s (IEA’s) [extended world energy
+balances](https://www.iea.org/data-and-statistics/data-product/world-energy-balances)
+(EWEBs) (Ayres, Ayres, and Warr 2003; Serrenho et al. 2014; Brockway et
+al. 2014, 2015). Data were stored in varying and inconsistent formats.
+An early SEA database (De Stercke 2014) estimated energy efficiencies of
+end-use machines using the economic status of countries, thereby
+precluding use in energy-economy studies.
+
+The authors of the current paper and others in the field wanted to
+expand SEA to cover all countries, but the spreadsheet approach to SEA
+was deemed not scalable. A new approach to SEA was needed, one that
+scaled across all countries for many years without relying on economic
+data to estimate machine efficiencies.
 
 ## CL-PFU database
 
 We created a suite of `R` packages to support the creation and use of
-the CL-PFU database. The CL-PFU database is a new resouce for the
-Societal Exergy Analysis (SEA) research community and assists SEA
-practitioners to analyze energy conversion chains (ECCs). The new
-packages enable, for the first time, scalable SEA.
+the CL-PFU database. The CL-PFU database is a new resouce for the SEA
+research community and assists SEA practitioners to analyze energy
+conversion chains (ECCs). The new packages enable, for the first time,
+scalable SEA.
 
 ## `R` packages
 
@@ -50,12 +69,54 @@ The packages in the following table are available on GitHub.
 
 <div id="refs" class="references csl-bib-body hanging-indent">
 
+<div id="ref-Ayres:2003ec" class="csl-entry">
+
+Ayres, Robert U., Leslie W. Ayres, and Benjamin Warr. 2003. “Exergy,
+Power, and Work in the US Economy, 1900–1998.” *Energy* 28 (3): 219–73.
+<https://doi.org/10.1016/s0360-5442(02)00089-0>.
+
+</div>
+
+<div id="ref-Brockway:2014aa" class="csl-entry">
+
+Brockway, Paul E., John R. Barrett, Timothy J. Foxon, and Julia K.
+Steinberger. 2014. “Divergence of Trends in US and UK Aggregate Exergy
+Efficiencies 1960–2010.” *Environmental Science & Technology* 48:
+9874–81. <https://doi.org/10.1021/es501217t>.
+
+</div>
+
+<div id="ref-Brockway:2015aa" class="csl-entry">
+
+Brockway, Paul E., Julia K. Steinberger, John R. Barrett, and Timothy J.
+Foxon. 2015. “Understanding China’s Past and Future Energy Demand: An
+Exergy Efficiency and Decomposition Analysis.” *Applied Energy* 155:
+892–903. <https://doi.org/10.1016/j.apenergy.2015.05.082>.
+
+</div>
+
+<div id="ref-De-Stercke:2014" class="csl-entry">
+
+De Stercke, Simon. 2014. “Dynamics of Energy Systems: A Useful
+Perspective.” Interim Report IR-14-013. Laxenburg, Austria: IIASA.
+
+</div>
+
[…]
