+++ ExampleFolder/OutputData/PipelineReleases/completed_allocation_tables/20231211T161340Z-1264f/data.txt
@@ -1,0 +1,10 @@
+file: completed_allocation_tables.rds
+file_size: 6937
+pin_hash: 1264f5e6628cf93e
+type: rds
+title: 'completed_allocation_tables: a pinned 713 x 14 data frame'
+description: ~
+tags: ~
+urls: ~
+created: 20231211T161340Z
+api_version: 1
