+++ ExampleFolder/OutputData/PipelineReleases/agg_eta_pfu_csv/20230817T180344Z-d424b/data.txt
@@ -1,9 +1,0 @@
-file: agg_eta_pfu_csv.csv
-file_size: 79797
-pin_hash: d424b5bdf399f38d
-type: csv
-title: 'agg_eta_pfu_csv: a pinned 720 x 12 data frame'
-description: ~
-tags: ~
-created: 20230817T180344Z
-api_version: 1
