+++ ExampleFolder/OutputData/PipelineReleases/eta_i/20231211T161342Z-f3eac/data.txt
@@ -1,0 +1,10 @@
+file: eta_i.rds
+file_size: 5767
+pin_hash: f3eac13a7d95b4b6
+type: rds
+title: 'eta_i: a pinned 48 x 7 data frame'
+description: ~
+tags: ~
+urls: ~
+created: 20231211T161342Z
+api_version: 1
