+++ paper.Rmd
@@ -129,6 +129,7 @@ See any number of internet references for an exergy primer,
 including @SciDirect:2023.].
 
 We created a suite of open-source `R` packages 
+and the metapackage `CLPFUDatabase`
 to assist SEA practitioners to analyze ECCs. 
 The new packages enable analysis of any country in the world
 across timespans of decades or longer.
