### 2023-12-19 9f5beb1
Updated paper.Rmd
