+++ ExampleFolder/OutputData/PipelineReleases/psut/20231211T161342Z-3e2fb/data.txt
@@ -1,10 +1,0 @@
-file: psut.rds
-file_size: 89917
-pin_hash: 3e2fb9f93811839b
-type: rds
-title: 'psut: a pinned 48 x 14 data frame'
-description: ~
-tags: ~
-urls: ~
-created: 20231211T161342Z
-api_version: 1
