+++ ExampleFolder/OutputData/PipelineReleases/psut_without_neu/20231211T161342Z-bf294/data.txt
@@ -1,10 +1,0 @@
-file: psut_without_neu.rds
-file_size: 108080
-pin_hash: bf29495f9b922a10
-type: rds
-title: 'psut_without_neu: a pinned 48 x 15 data frame'
-description: ~
-tags: ~
-urls: ~
-created: 20231211T161342Z
-api_version: 1
