### 2023-12-19 fdf813c
Add new pins.
