+++ ExampleFolder/OutputData/PipelineReleases/eta_i/20231219T204417Z-e0fd6/data.txt
@@ -1,0 +1,10 @@
+file: eta_i.rds
+file_size: 5870
+pin_hash: e0fd679180333603
+type: rds
+title: 'eta_i: a pinned 48 x 7 data frame'
+description: ~
+tags: ~
+urls: ~
+created: 20231219T204417Z
+api_version: 1
