+++ ExampleFolder/OutputData/PipelineReleases/psut/20231219T204403Z-2c3ce/data.txt
@@ -1,0 +1,10 @@
+file: psut.rds
+file_size: 91261
+pin_hash: 2c3ce5a40f3815c1
+type: rds
+title: 'psut: a pinned 48 x 14 data frame'
+description: ~
+tags: ~
+urls: ~
+created: 20231219T204403Z
+api_version: 1
