+++ paper.Rmd
@@ -355,6 +355,7 @@ devtools::install_github("EnergyEconomyDecoupling/CLPFUDatabase")
 #     Sample data for two countries (Ghana and South Africa) 
 #     and two years (1971 and 2000) are included (with permission)
 #     at the correct location in the repository for this paper.
+
 # (2) Download FAO and ILO data for muscle work calculations
 #     via the included script.
 #     This download step is needed only once, and
