### 2023-12-20 e49757f
Eliminate date field.
