+++ paper.Rmd
@@ -21,7 +21,6 @@ affiliations:
   index: 1
 - name: School of Earth and Environment, University of Leeds
   index: 2
-# date: "17 Aug 2023"
 bibliography: paper.bib
 output:
   html_document:
