### 2023-12-20 2660644
Added ERE paper reference.
