+++ paper.bib
@@ -1,12 +1,20 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Matt at 2023-08-08 10:50:25 -0400 
+%% Created for Matt at 2023-12-20 14:07:38 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@article{Brockway:2024aa,
+	author = {Paul E. Brockway and Matthew Kuperus Heun and Zeke Marshall and Emmanuel Aramendia and Paul Steenwyk and Thomas Relph and Michelle Widjanarko and Jeonghoo (James) Kim and Anjana Sainju and Julian Irtube},
+	date-added = {2023-12-20 14:03:06 -0500},
+	date-modified = {2023-12-20 14:07:25 -0500},
+	journal = {Environmental Research: Energy (in review)},
+	title = {A Country-Level Primary-Final-Useful ({CL-PFU}) Energy and Exergy Database: Overview of its construction and 1971--2020 world-level efficiency results},
+	year = {2024}}
 
 @manual{Aramendia:2022ab,
 	author = {Emmanuel Aramendia and Matthew Kuperus Heun},
@@ -15,7 +23,8 @@
 	note = {R package version 0.1.1},
 	title = {EROITools: Tools for calculating and aggregating Energy Return On Investment values},
 	url = {https://doi.org/10.5281/zenodo.7696869},
-	year = {2022}}
+	year = {2022},
+	bdsk-url-1 = {https://doi.org/10.5281/zenodo.7696869}}
 
 @manual{Aramendia:2022aa,
 	author = {Emmanuel Aramendia and Matthew Kuperus Heun},
@@ -24,7 +33,8 @@
 	note = {R package version 0.1.5},
 	title = {ECCTools: Tools for modifying the Energy Conversion Chain},
 	url = {https://doi.org/10.5281/zenodo.5841962},
-	year = {2022}}
+	year = {2022},
+	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5841962}}
 
 @phdthesis{Aramendia:2023aa,
 	author = {Emmanuel Aramendia and Paul Brockway and Peter Taylor and J. Norman and Matthew Kuperus Heun and Zeke Marshall},
@@ -33,7 +43,6 @@
 	school = {University of Leeds},
 	title = {Chapter 4: Estimation of fossil fuels useful stage energy return on investment and implications for renewable energy systems \normalfont in {E}xploring the useful energy implications of the global energy transition: a net energy perspective},
 	year = {2023}}
-
 
 @misc{International-Energy-Agency:2023aa,
 	author = {{International Energy Agency}},
@@ -44,62 +53,61 @@
 	title = {{IEA} {W}orld {E}xtended {E}nergy {B}alances},
 	year = {2023}}
 
-
 @article{Aramendia:2022tv,
 	author = {Emmanuel Aramendia and Matthew Kuperus Heun and Paul E. Brockway and Peter G. Taylor},
+	doi = {10.1016/j.apenergy.2021.118413},
 	journal = {Applied Energy},
 	month = {February},
 	number = {118413},
 	pages = {1-20},
 	title = {Developing a Multi-Regional Physical Supply Use Table framework to improve the accuracy and reliability of energy analysis},
 	volume = {310},
-	year = {2022}, 
-	doi = {10.1016/j.apenergy.2021.118413}}
-
+	year = {2022},
[…]
