+++ paper.bib
@@ -215,10 +215,10 @@
 	year = {2023},
 	bdsk-url-1 = {https://doi.org/10.5281/zenodo.8226419}}
 
-@manual{Heun-PFUDatabase:2023,
+@manual{Heun-PFUPipeline:2023,
 	author = {Matthew Kuperus Heun and Zeke Marshall},
 	note = {R package version 0.1.1},
-	title = {PFUDatabase: Functions and Workflows for Primary-Final-Useful Societal Exergy Analysis Within the PSUT Framework},
+	title = {PFUPipeline: Functions and Workflows for Primary-Final-Useful Societal Exergy Analysis Within the PSUT Framework},
 	url = {https://doi.org/10.5281/zenodo.5228375},
 	year = {2023},
 	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5228375}}
