+++ paper.bib
@@ -223,10 +223,10 @@
 	year = {2023},
 	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5228375}}
 
-@manual{Heun-PFUAggDatabase:2023,
+@manual{Heun-PFUAggPipeline:2023,
 	author = {Matthew Kuperus Heun},
 	note = {R package version 0.1.2},
-	title = {PFUAggDatabase: Calculate Aggregations and Efficiencies For Energy Conversion Chains in PSUT Format},
+	title = {PFUAggPipeline: Calculate Aggregations and Efficiencies For Energy Conversion Chains in PSUT Format},
 	url = {https://doi.org/10.5281/zenodo.6409759},
 	year = {2023},
 	bdsk-url-1 = {https://doi.org/10.5281/zenodo.6409759}}
