### 2023-12-20 75cf493
Update bibliography
