+++ paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Matt at 2023-12-20 16:40:00 -0500 
+%% Created for Matt at 2023-12-20 16:42:51 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -11,10 +11,10 @@
 @article{Brockway:2024aa,
 	author = {Paul E. Brockway and Matthew Kuperus Heun and Zeke Marshall and Emmanuel Aramendia and Paul Steenwyk and Thomas Relph and Michelle Widjanarko and Jeonghoo (James) Kim and Anjana Sainju and Julian Irtube},
 	date-added = {2023-12-20 14:03:06 -0500},
-	date-modified = {2023-12-20 16:40:00 -0500},
+	date-modified = {2023-12-20 16:42:51 -0500},
 	journal = {Environmental Research: Energy},
+	pages = {(In review)},
 	title = {A Country-Level Primary-Final-Useful ({CL-PFU}) Energy and Exergy Database: Overview of its construction and 1971--2020 world-level efficiency results},
-	volume = {(In review)},
 	year = {2024}}
 
 @manual{Aramendia:2022ab,
