+++ paper.Rmd
@@ -345,7 +345,6 @@ following the example below.
 # Example
 
 ```{r, eval=FALSE}
-library(devtools)
 # (0) Use the CLPFUDatabase metapackage to install all dependencies.
 #     This package installation step is needed only once.
 devtools::install_github("EnergyEconomyDecoupling/CLPFUDatabase")
