+++ paper.bib
@@ -1,12 +1,22 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Matt at 2023-12-20 16:42:51 -0500 
+%% Created for Matt at 2023-12-21 11:05:54 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@manual{Heun-RCLabels:2023,
+	author = {Matthew Kuperus Heun},
+	date-added = {2023-12-21 11:04:57 -0500},
+	date-modified = {2023-12-21 11:05:39 -0500},
+	note = {R package version 0.1.5},
+	title = {RCLabels: Tools for munging and manipulating row and column labels on matrices},
+	url = {https://doi.org/10.5281/zenodo.5819143},
+	year = {2023},
+	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5819143}}
 
 @article{Brockway:2024aa,
 	author = {Paul E. Brockway and Matthew Kuperus Heun and Zeke Marshall and Emmanuel Aramendia and Paul Steenwyk and Thomas Relph and Michelle Widjanarko and Jeonghoo (James) Kim and Anjana Sainju and Julian Irtube},
@@ -151,13 +161,14 @@
 	year = {2018},
 	bdsk-url-1 = {https://doi.org/10.1016/j.apenergy.2018.05.109}}
 
-@manual{Heun-RCLabels:2023,
+@manual{Heun-CLPFUDatabase:2023,
 	author = {Matthew Kuperus Heun},
+	date-modified = {2023-12-21 11:05:12 -0500},
 	note = {R package version 0.1.5},
-	title = {RCLabels: Tools for munging and manipulating row and column labels on matrices},
-	url = {https://doi.org/10.5281/zenodo.5819143},
+	title = {CLPFUDatabase: Easily Install and Load Packages to Construct the CL-PFU Database},
+	url = {https://doi.org/10.5281/zenodo.10359832},
 	year = {2023},
-	bdsk-url-1 = {https://doi.org/10.5281/zenodo.5819143}}
+	bdsk-url-1 = {https://doi.org/10.5281/zenodo.10359832}}
 
 @manual{Heun-matsbyname:2023,
 	author = {Matthew Kuperus Heun},
