### 2023-12-21 #6057 @MatthewHeun
@timtroendle:

I believe I have addressed all of the issues raised by the reviewers at https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper/issues. The review process can proceed at your convenience.

@nmstreethran, @abhishektiwari:

Thank you for your constructive and helpful comments! They have improved both the paper and the software.

Happy holidays, everyone!
