### 2024-01-03 #6057 @timtroendle
@MatthewHeun, could you have a look at the two problems above? First, there is an invalid DOI in your reference list. Second, your new affiliations are causing a technical issue. I suppose quoting them should resolve the issue.

Once that's done, I will have a final look at the manuscript and we can move forward.

In addition, I believe you have not yet archived the meta-package anywhere, correct? Could you please
- [ ] Make a tagged release of the metapackage, **and** list the version tag of the archived version here.
- [ ] Archive the reviewed software **in** Zenodo **or** a similar service (e.g., figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., **in** Zenodo) has the correct metadata. This includes the title (should match the paper title) **and** author list (make sure the list **is** correct **and** people who only made a small fix are **not** on it). You may also add the authors’ ORCID.
