### 2024-01-03 #6057 @MatthewHeun
@timtroendle, I'm happy to dig into those three problems. I'll add them as separate issues to the paper's repository at https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper/issues, so you can track progress. I'll also ping you again here when everything is done.
