### 2024-01-03 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#19 @MatthewHeun
Archive CLPFUDatabase metapackage
@timtroendle requested that I do the following:

- [x] Make a tagged release of the metapackage (CLPFUDatabase), and list the version tag of the archived version at https://github.com/openjournals/joss-reviews/issues/6057.
- [x] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors’ ORCID.
- [x] Please list the DOI of the archived version at https://github.com/openjournals/joss-reviews/issues/6057.
