### 2024-01-03 EnergyEconomyDecoupling/CL-PFU-JOSS-Paper#19 @MatthewHeun
I made a tagged release of metapackage `CLPFUDatabase`. The DOI is 10.5281/zenodo.10456048.

Zenodo does not allow ":" characters in the `title:` field of the `CITATION.cff` file. So, the title in the `CITATION.cff` file is different from the title of the paper:

`CITATION.cff`: CLPFUDatabase A suite of R packages for energy conversion chain analysis
Paper: CLPFUDatabase: A suite of R packages for energy conversion chain analysis

Otherwise, I believe all requests have been addressed. Please let me know if you have any questions.
