### 2024-01-03 #6057 @MatthewHeun
@timtroendle: I just made a tagged release of the metapackage. See https://github.com/EnergyEconomyDecoupling/CL-PFU-JOSS-Paper/issues/19 for details. The version is v0.1.8. The DOI is 10.5281/zenodo.10456048 (https://doi.org/10.5281/zenodo.10456048)

Please let me know if you have any questions.
