### 2024-01-04 #6057 @timtroendle
@MatthewHeun it looks all good to me. Apart from the invalid DOI, there is one more tiny issue to fix: the author list on Zenodo and in the paper differ. Can you please make sure both lists have the same order?
