### 2024-01-04 #6057 @MatthewHeun
@timtroendle Thanks for your eagle-eyed spot of the different author order. That problem has now been fixed. Unfortunately, that meant I needed to generate a new release of the `CLPFUDatabase` package. I hope that doesn't mess up your system.

- New release: v0.1.9
- New doi: 10.5281/zenodo.10458774 (https://doi.org/10.5281/zenodo.10458774)

We're working on the invalid DOI, which is the input data repository for the database. We hope to have that done by early next week. So please stay tuned.
