### 2024-01-08 #6057 @MatthewHeun
@timtroendle Thanks for your patience as we sorted out the un-minted DOI. The un-minted DOI contains output products from the database and is intended for a paper in review at the journal Environmental Research: Energy. We decided to remove the references to the un-minted DOI from the JOSS paper. The JOSS paper points to the Environmental Research: Energy paper which contains a references to the un-minted DOI. The un-minted DOI will go live at a later date coinciding with the publication of the Environmental Research: Energy paper.

I think this means we can move forward with the publication of the JOSS paper now.

Thanks for your deft guidance of this paper through the review process at JOSS. Also, my sincere thanks to @nmstreethran and @abhishektiwari for their reviews!
