### 2024-01-08 cfcea8f
Remove invalid DOI from .bib file.
