+++ paper.bib
@@ -254,14 +254,6 @@
 	year = {2021},
 	bdsk-url-1 = {https://doi.org/10.21105/joss.02959}}
 
-@dataset{Marshall:2023aa,
-	author = {Zeke Marshall and Matthew Kuperus Heun and Paul E. Brockway and Emmanuel Aramendia and Paul Steenwyk and Thomas Relph and Michelle Widjanarko and James (Jeonghoo) Kim and Anjana Sainju and Julian Iturbe Franzius},
-	doi = {10.5518/1199},
-	publisher = {University of Leeds},
-	title = {A Multi-Regional Primary-Final-Useful (MR-PFU) energy and exergy database v1.0.},
-	year = {2023},
-	bdsk-url-1 = {https://doi.org/10.5518/1199}}
-
 @book{Rocco:2016,
 	address = {Switzerland},
 	author = {Matteo Vincenzo Rocco},
