### 2024-01-08 #6057 @MatthewHeun
> Hm the invalid DOI still appears in the reference check, but it's not included in the paper and so this looks good to me.

@timtroendle The invalid DOI was still in the .bib file but not referenced in the paper. I just removed the invalid DOI from the .bib file and pushed. So if you run another check, it should look OK.

We will look at the paper and get back to you with a final word by end of day tomorrow.

Thanks again for your good work!
