### 2024-01-09 #6057 @MatthewHeun
@kyleniemeyer We're done. Only four minor changes:

- inserted "and" for readability,
- moved a reference to end of sentence instead of beginning,
- capitalized Extended World Energy Balances, and
- added a missing reference to the PUFPipeline package).

Feel free to move the article to the next step of the publication process. We won't make any further changes.
