### 2025-07-15 #6057 @paulrougieux
Dear @MatthewHeun your paper has been marked as a similar JOSS papers, and the process suggested to "click through to the review issue for that paper and invite one or more of the authors". Would you like to review my paper [Cobwood: Enhancing Forest Economics Model Reusability Through Labeled Panel Data Structures](https://github.com/openjournals/joss-reviews/issues/8587#issuecomment-3071248774) ?
